Use jemalloc for long-running, high-churn services
We've noticed poor-memory reuse in iot_config, particularly when
updating regions. There is a step-change in memory utilization when
regenerating region params, as can be expected, but memory goes up a
little bit on subsequent regens, when in theory there should be enough
deallocated memory laying around to handle the needs.