Don't allow constructing a RouteId with invalid uuid
Making sure a user provides a valid Uuid as early as possible can
provide better guidance towards using the system. Normally the route id
is not attempted to be parsed until the db is about to be hit where
other concerns are more pertinent. This leaves Uuid parsing errors
mostly to be overwritten as general authentication errors.