tighten gw info perm check to require proto request
The method before was named specifically for the info request, but the
function signature was much looser than the name implied.

Passing the request reduces the chance of messing up argument order, or
misunderstanding the arguments.

If at some point more methods require the ability for self
authorization, hopefully a better abstraction will reveal itself.