Landtype Boosting (#789)
* add landtype to hex boosting

- initialize hexes without a landtype assignment
- load landtype disktree
- update tests to match HIP-103 chart with all 3 data sets

* use struct HexAssignments

HexAssignemnts wrap and labels Assignments. The order Assignments are
considered is not always clear unless you have HIP-103 up next to the
code. So we try to funnel all uses of the fields together into one
place, and callers deal with Assignments as a collection.

* reify HexBoostData for CoverageDaemon

This is an attempt to simplify the trait soup hex boosting was starting
to become. Until there is a concrete need for more traits, I believe
this flow is easier to follow, and easier to mock at the level desired
in tests.

* remove unused trait

* add mobile-verifier command to verify disktree

This command makes sure all values in a disktree file can be turned into
Assignment.

* rename WorldCover -> Landtype to match HIP

* add from_paths constructor to Geofence

Now a geofence can be constructed directly without needing to go through
a file if you have a HexTreeSet.

* match printing order with HIP order

The tables in HIP-103 list the assignments (AAA, ABC, etc,.) in the
order (footfall, landtype, urbanized). Having the Debug repr of
Assignments print in the same order makes it easier to cross reference
with the tables.

* Unit test for HexAssignments

This test will break if the logic for deriving scores changes. If the
break is on purpose, the test should be updated.

* match comments

* update proto dep after rebase

* Update number of fields in hex update query

* remove unused test code

* implement From for the proto type, avoid base type

This forces us to go through the proto enum, but provides a stronger
contract for going between types. very nice

* update proto back to master

---------

Co-authored-by: Brian Balser <bbalser@nova-labs.com>