Add device_type to BoostedHexInfo from db
The device_type column is assumed to be a jsonb, because the device_type
column for the `mobile_hotspot_infos` table is an enum, and a jsonb
field.

It's nullable, which will represent boosting a hex for ALL device types.

Otherwise, a camelcase value mapping to the DeviceTypeV0 in
helium-program-library.
https://github.com/helium/helium-program-library/blob/master/programs/hexboosting/src/state.rs

This is different from the snake_case values used with protobufs.