Add test for accumulating hex boosts
This test uses a single type of radio to reduce adding noise to rewards
by factors other than hex boosting.

The first pass of the test calculates rewards for 2 radios with no
boosted hexes to prove they are in equal standing.

The second pass boosts only the hex for single radio, but with multiple
BoostedHexDeviceType's to ensure they accumulate.