make boosted hex test function more explicit
If the expired check had been made a global check, the ability to use
BoostedHexes for modified hexes would have broken at runtime.

The attempt here is to make very explicit during testing how to meet the
same contract as the database queries for boosted hexes. I think there
are still some cracks, but we can narrow in on those as we find them.
For now, I think naming test constructor functions is a good start.