Mobile Config filter Boosted Hexes (#801)
* Refactor to use BoostedHexes methods

Making the internal member hexes private will make it easier to change
the implementation when device type is introduced.

* Add device type to boosted hex info

* refactor metadata_db tests to make test clearer

also makes it easier to add new tests

* remove expired boosted hexes when streaming from db

* ensure no tests are written with expired boosted hexes

* optimize by computing end_ts in db query

Thanks for the query help Brian!

By precomputing the end timestamp of a boosted hex, we can not have to
stream all the hexes out of the db just to throw them away.

* fixup after rebase

- remove unused imports
- remove old refactor return types

* make boosted hex test function more explicit

If the expired check had been made a global check, the ability to use
BoostedHexes for modified hexes would have broken at runtime.

The attempt here is to make very explicit during testing how to meet the
same contract as the database queries for boosted hexes. I think there
are still some cracks, but we can narrow in on those as we find them.
For now, I think naming test constructor functions is a good start.