Make calculating coverage points a top level function
I want this crate to make the algorithm one of the most important things. Hiding it inside a structs impl block seems not a great of doing that, I think. I also don't think the rewardable radio impl block should be moved to the top to make the algorithm more front and center, in my mind it would only make RewardableRadio look more important.