Location trust scores
trust scores and covered hexes are linked in many ways. I'm opting to precompute multiple values then have quick access to them, rather than waiting until they're needed and computing them for every hex. Or carrying around a spot for computed data that get's filled on the first use.

When the rest of the types are fleshed out, some benchmarks will be done to determine if this is a problem.