Take responsibility of rounding shares and total coverage points
For consistency, the calculator is now truncating values that would have been truncated be the user of the calculator, so we avoid a situation where 2 users of the calculator are using different rounding strategies. 

The fields that make up the truncated values are provided untouched.

Truncating fully to a u64 causes calculations to be off by quite a margin.