hip-103: provider boost increases oracle boost always
If a hex is boosted by a provider, that hex is deemed valuable from an oracle context, and should always be 1x.

Hexes were being "cleaned" of their boost values _before_ trying to determine the oracle multplier, which coupled a hexes boost effect to the assignment multiplier to wether or not the radio was eligible for boosting rewards. 

That was incorrect. Hexes are now "cleaned" at the same time as determining the multipliers.