move collections to the end
I could see an argument for the collection being moved to the front as well. But I think either is a better case than in the middle.

I chose the end here because I think it reads easier when testing to have the function, then small decision making structs, _then_ the collection that could be constructed in place.