clean up trust_score construction and long types
Move zipping trust scores with their distances to heartbeatreward.
Leave

I put the use statements above where they are in the function because the names of the types kind of overlap with the names used in the mobile verifier, and I wanted to make it extra clear to anyone reading, this function is transforming those values into calculator types.