construct coverage map inside block scope
We can insert coverage objects directly after iterating over them, radio_info only needs to know if the radio `is_indoor`. Moving that binding to the top helps hint that it's set by something in the block and returned.

And we don't need to carry around covered_hexes for any longer than necessary.