Rename trait FileStoreAsBytes -> MsgBytes to follow convention for prost traits
To keep from dealing with Vec<u8> and String auto implementations for
prost::Message that disallows using a file sync with anything but a
prost::Message.

We're going to follow the other Msg* trait conventions where we
implement them only for the types we are using.

This allows someone to make a file sync for whatever type they may have,
and convenience for proto types we write reports about.