rename file sink trait to writer
The abstraction comes in two parts.

The first we can get a file sink from a type if it implements the
FileSinkWrite trait.

The second is a type being able to be written to a file sink if it can
be turned into bytes.

This separates the convenience from the actual functionality.