update hex boosting tests for radio_reward_v2
It becomes more difficult to compare ratios with radio_reward_v2 because
boosted rewards are scaled. In many of the test we were hoping to have
nice ratios (1:10) of poc_rewards. However, this was mostly because a
radio would have received 0 boosted_rewards. Checking that ratio is
impossible.

We can however check the coverage_points (now that they mean points, and
not rewards) for the expected ratio, and rely on other tests to ensure
we're calculating poc_reward amount correctly.