Add Utility struct NoState for FileInfoPoller
Sometimes, managing the state of processed files is more cumbersome than
just reading back further in time, or writing a new file.

This came along with the Service Provider Funds reports that for a long
time will have a single item in them, and it didn't seem worth the
overhead to add a database to a service that could simply upload a new file.