return bps directly
because the promotion_fund workspace is writing out a proto report,
fetching from solana should return the actual value on chain, so there
is no chance it get's changed during conversions to and from percents.

This does mean the user of this value needs to know to change it to a
percentage at the time of use.