Fallback to confirming transactions on all errors
If a submitted transaction does not come back as a defacto success, we spawn a process that checks a txn made it to the chain by signature. By default, we will retry 5 times a minute apart, before considering the txn an actual failure to be tried again later.

If during the checking we find the txn on chain, we do the same logic as if it had passed the first time.