HIP-134 unique connections (#904)
* point to unique connections proto

* add ingest for unique connections msg

* verify unique conection reports

* start db queries for unique connections

* add UniqueConnectionCounts to CoverageShares

* Update enum name for doc links

* calculate qualified radios oracle boost multiplier

* add test for qualified wifi in bad coverage areas

* add poc and dc rewarder test for qualified wifi hotspots

* spell out received_timestamp in db like other tables

* clear unique connections after rewarding

* make unused struct private

* Move unique connections to its own namespace

while it currently comes from the radio-threshold service, and it feelslike a threshold for a radio to pas for rewarding; current radio thresholds is about service provider boosting, and unique connections is about oracle boosting.

* move unique connections check to mod

* add UniqueConnectionIngestor to mobile-verifier server

* clippy updates

* add banned radio to qualified radio test

* make sure qualified radios are exempt from sp banning

* Add primary key for hotspot_pubkey and received_timestamp

this makes sure we're idempotent when reprocessing reports within the same day.

* Add test to make sure we always grab the latest value from the database for a record

If we get sent a new report in the same day, we always want to use the latest supplied number.

* Add conflict for saving, distinct for getting

If we get sent a new report in the same day, we always want to use the latest supplied number.

* update proto to include gateway metadata v2

* Fix distinct on query

The old distinct on was returning all rows for a pubkey in reverse
order, so the latest value was always inserted into the hashmap last.

Now, we are correctly getting a single value for a pubkey that is the
latest by timestamp.

Thank you Brian.

* fix be inclusive on start and exclusive on end

If we use inclusive on the end date, we may use a previous days record
for the current day when we do not mean to.

* batch insert verified reports

batch inserting is more performant from the IO perspective.

* update proto, add oracle boost hex status to reward_v2

send oracle boosting status in reward_v2. sp boosted status has been
renamed to reduce confusion. orable boost status has been added

* add unique connections check to run rewarder

* add tests for finding reports to continue rewarder

* go back to proto master