write out data transfer sessions for pending burns
When we submit to solana, we move the existing data transfer sessions to a pending table. When the txn is finalized, we write those to s3 from the pending table and delete.

We do this move so the Daemon can continue to ingest s3 files without making them indistinguishable from pending txns.