Helium Lib attempt: 2 (#930)
* use local helium-lib for testing

* Remove SolanaRpcError::Test

* misformatted tokio::select!

* Use helium-lib for burn txns

* Initialize SolanaRpc with SubDao

* update helium-lib deps

* Use helium-lib for sending and confirming mobile burn txn

* use helium-lib for sending and confirming iot burn txns

* remove panics

* Update to receive TransactionWithBlockhash and Name

* Start updating integration tests to use new helium-lib interactino, and an actual db

* Use sqlx for iot integration tests

Remove test mocks, or move them to unit tests where applicable

* Box potential large error

* remove unneccessary clone

* Add inner impl for adding pending transaction to override date

now we can remove the ignore from testing confirmed txns

* remove duplicated integration test with a bunch of mocks

* Move send_with_retry macro into start_boost

It's the only place it's still used, and a sender is being worked on that will use the same name.

* Bring in TxnSender from helium-lib to Oracles

The work done is helium-lib is useful, but not quite what we were looking for.

The change still allows us to clean up quite a bit within Oracles though, so we can still use it.

* Remove retry options

* Update logging

* Move TestSolanaClient to test file

* Point at published helium-lib branch

* Remove unused callbacks

* Bring in sender unit tests

* more reasonable real world default, and override for test

* Go all the way back to helium-lib master

* Use local version of wrapping transaction

* Unsplit error conditions

* Rename so both stores have the same name

* Alias SolanaTransaction so we don't have ot use a qualified path everywhere

* Start adding integration tests for mobile packet verifier

* check balances and written reports

* Track pending txns for mobile burns

* confirm pending txns on startup

* Remove unused struct

* Move solana test client to solana workspace

both packet verifiers share a test impl

* I think we're past this now

* fixup comments

* use helium-lib for making boosting txn

* Remove unused noopstore

* local noopstore

* return full error from preparation

* confirm_pending_txns takes a FileSinkClient

* test for confirm pending txns writing out data transfer sessions

* write out data transfer sessions for pending burns

When we submit to solana, we move the existing data transfer sessions to a pending table. When the txn is finalized, we write those to s3 from the pending table and delete.

We do this move so the Daemon can continue to ingest s3 files without making them indistinguishable from pending txns.

* clippy fixes

- unneeded lifetimes
- unnecessary references
- box large Error

* A txn that cannot be finalized will move back to available

* Break out test saving multiple data transfer sessions

* Helper TestBurner

* add logging around burn success/failure

* Move txn db function to new crate

* get_payer_balance helper

* transcation -> txn

* move bytes_to_dc to lib

* clippy suggestions

* Add ability to panic solana client when sending transaction

* Skip burning when there is a pending txn

* Test for skipping burn when there are pending txns

* clippy updates

* always confirm before burn, don't confirm on startup

we want to put the server in a place to write out verified data transfer sessions, even if it cannot burn.
fixup to always confirm before burning

* Use the same function for saving data transfer sessions in all places

This is a relatively big query, and it is nice to not have to keep multiple copies of it in sync.

first_timestamp was added to handle the case where pending data transfer sessions are being moved back to the regular data transfer sessions table.