`helium-lib` update (#997)
`helium-program-library` had to update to `anchor@>30`.
That version of `anchor` broke backwards compatibility with the IDL format, so we had no choice but to move forwards.

This meant, updating `helium-lib` which was a desire of ours anyway.
And it finally brought with it some boons.

- drop `helium-anchor-gen` as a dep, all programs can be accessed through `helium-lib`
- update `sqlx`, in fact it was required.

Updating `sqlx` has been a pain for us in the past because a bunch of anchor/solana dependencies wreaked havoc with very low level dependencies like `zeroize`. Now that we've updated, those dependencies are sequestered enough in `helium-lib` that we can unpin from our special version of slqx and go back to the main branch.

But wait! Another dragon appears to have been hiding behind the first dragon.

We udpated to `sqlx@0.7`, but we need `sqlx@0.8` for this dragon. There were a few bugs in the very short-lived `sqlx@0.7`. So @jeff did the needful and pushed us through to `0.8`. 

Below is a truncated list of the struggles to get us to the mountain top.

* update bs58

* Moved into solana workspace 
  * anchor-client
  * helium-anchor-gen (just you wait)
  * solana-client
  * solana-sdk
  * solana-program
  * spl-token
  * helium-lib

* upgrading sqlx and db connection type errors
  * fixup iot-packet-verifier
  * fixup iot-verifier
  * fixup mobile-config
  * fixup reward-index
  * fixup mobile-packet-verifier
  * fixup moile-verifier
  * fixup boost-manager
  * fixup iot-config

* Go back to some updated deps that didn't need to be
  * back to known backon
  * back to known metrics

* Some code cleanliness
  * order imports for boost-manager
  * Export SolPubkey as top level
  * Export RpcClient as top level for price
  * Remove solana from iot-config
  * reorder cargo for iot-packet-verifier
  * Cleanup imports for iot-verifier
  * update to SolPubkeyin mobile-config
  * Update to top level exports for mobile-verifier
  * export SubDao as top level
  * merge imports for helium-lib
  * format cargo.toml files

* xorf-generator and helium-lib needed some more updates to work with Oracles
  * back to main branch for helium-lib and xorf-generator

* Jeff has entered the dance and is boogying 
  * update sqlx to 0.8
  * disable default sqlx features
  * roll back majority of branch deps from helium-crypto
  * restore helium-lib to main branch

---------

Co-authored-by: jeffgrunewald <jeff.grunewald@gmail.com>