make FileInfoPollerParser non-async trait
Simplify making a Parser by removing the need to know how to deal with a stream, and only have to deal with decoding a single message at a time.

The only required function to implement is `decode_item`, the rest can handle itself.

Default implementations are provided for hooking into custom logging for stream and decode errors. Or if you want to handle the whole stream by yourself.

These methods are instance methods to allow for stateful parsers.