Add metric when we throttle route stream updates
Broadcast channels will drop messages if one of the consumers falls 
behind. We cannot miss updates for downstream HPR. This is why we check 
how full the broadcast channel is before queueing a message. 

Simply sleeping before adding a message means it’s theoretically 
possible for messages to be queued out of order. 

Before we spend resources trying to tackle this potential case, we 
should add a way to check how common it might be.