remove load timeout settings from filestore
It has been removed in updated version of aws sdks, there are ways to
handle it, but they're much more involved.

We can test with updated versions of loading credentials, and it we're
still getting timeout failures for price tracker, we can consider adding
it back in with the new style.