Update how we calculate our start time for file info poller
For starting a service fresh, we normally fill in a `start_after` 
settings to ensure we don’t process all files from the beginning of time
if we don’t need to.

However, sometimes we want to move forward and skip processing some 
data. Upating the `start_after` setting would do nothing because if a 
timestamp was present in the database, it always took precedence.

This change makes it so we always choose the latest possible date. 

Normal operation is maintained where we start with a `start_after` 
timestamp around the time a service is hooked up, and we want to 
continue from the latest timestamp in the DB after that. But if the 
`start_after` setting is set to later than the time in the DB, we will 
use the setting instead. 

This allows us to update `start_after` and have it be respected without 
needing to manually drop rows from the `files_processed` table, or 
insert rows that are lies.