FileInfoPollerConfigBuilder add builder setter (#1059)
* FileInfoPollerConfigBuilder add builder setter

Allow any value that can be turned into a `LookbackBehavior` to be 
passed into the builder.

The interval values of `LookbackBehavior` do not overlap, so we can 
safely determine what kind of behavior we’re trying to achieve by 
passing a plain `DateTime<Utc>` or `Duration`.

* Add custom setters for lookback behaviors

Remove the implicit casting added by `#[builder(setter(into))]` in lieu 
of providing specific setters for lookback variants.

The new setters are prefixed with `lookback_` to make them discoverable 
when setting `lookback`, and it’s a nice place to hang some 
documentation.