Attemting to move FileWriteExt trait out of file-store into file-store-helium-proto
This operation is confusing because we run into a slew of rust orphan 
rule problems. We’re trying to keep the type we’re writing in 
FileSink<T>, but cannot tie it to a trait within file-store, because we 
don’t have enough leverage outside of file-store to implement it for 
anything.