Remove MsgBytes
It’s an extra layer of indirection that makes it hard to use other proto
types with file-store.

Originally MsgBytes had to be implemented for any type that wanted to 
use FileStoreWriteExt and every type had to have a manual implementation
cecause of our lack of ability to provide a blanket impl for 
prost::Message. Almost literally everything implements prost::Message.

We don’t lose any type safety because we still type our FileClients and 
the like with the protobuf type we expect to be writing.