Mj/file store helium proto (#1068)
* Remove csv from file-store

Nothing is reading/writing to csv in file-store. And the dump command 
follows the pattern of dumping json to stdout.

* Remove unused Shutdown error

* Remove unused tokio::JoinError

* return config::ConfigError from config

file-store proper doesn’t really do anything with settings (cli 
commands, but okay), to where we need to wrap a config error in a 
special library type.

We can return a config::ConfigError and the caller can map it how they 
need on their application startup.

* Cli Commands to return anyhow::Error

We don’t need a special library error from a cli. The internals can 
return file_store::Error, and we can turn them into anyhow. This is 
planning for being able to break the cli out into a proto specific 
workspace for helium proto types.
cli dump command to return anyhow::Result

and use MsgDecode where we can and it doesn’t confuse things

* Break out file-store cli commands

* Move all helium-proto files to new workspace

* Move print_json helper to cli workspace

* Fixup imports to use file_store_helium_proto

* Move traits around

- ReportId and IngestId to iot_verifier where they’re used solely
- TimestampEncode/Decode to file_store_shared

* Fix re-export of shared Error Result types

* Attemting to move FileWriteExt trait out of file-store into file-store-helium-proto

This operation is confusing because we run into a slew of rust orphan 
rule problems. We’re trying to keep the type we’re writing in 
FileSink<T>, but cannot tie it to a trait within file-store, because we 
don’t have enough leverage outside of file-store to implement it for 
anything.

* Accept any T for fileSink that implements prost::Message

This get’s rid of the need for MsgBytes, but keeps us typed over a proto
message that that can serialize itself to bytes.

* Remove FIleType from tests

* Make a more generic From impl for FileInfo

We can turn into FileInfo from anything that knows how to stringify 
itself and a DateTime.

* Move FileType enum from file-store to file-store-helium-proto

* Move file-store-shared back into file-store

Breaking out those shared components helped me think through a bunch of 
the changes that were made, but they don’t need to be their workspace 
anymore.
remove file-store-shared

* Move file-store-helium-cli into file-store-helium-proto

It was nice to have all that out of the way when working on refactors, 
but it’s not useful enough on its own to deserve a workspace.
remove file-store-helium-cli
remove more cli code

* Fixup file_store TimestampEncode/Decode trait imports

* Move MsgVerify trait to file-store-helium-proto

* Add metwork_common grouping, and move crates into mod.rs

* remove top level trait export

* file-store-helium-proto cli return anyhow::Result

The CLI is application layer, and does not need to constrain itself to a
file_store::Error.

* Bring back the sqlx-postgres feature for the err variant

* Bring back the proto roundtrip tests

* Remove not needed type aliases

* Bring back FileInfoPollerConfigBuilderError

doesn’t make sense to remove this one yet imho

* rename file-store-helium-protos -> file-store-oracles