Enforce bucket cleanup with local testing
It would be ideal to throw a compiler error in the drop implementation 
of AwsLocal that would only get included if the trait was used.

The best we can do right now is panic at the end of a test if the Drop 
trait is called automatically. 

https://doc.rust-lang.org/std/mem/struct.ManuallyDrop.html
bypasses the Drop trait.