Better implementation of drop gaurd
The previous version used `std::mem::ManuallyDrop` which achieved the 
goal.

However, it had the unintended side effect not dropping the internal 
`BucketClient`. Which we absolutely want to clean up.

This version is a bit simpler, it simply flips a toggle of wether or
not we want to panic when the test ends.