Enforce bucket cleanup with local testing (#1092)
* Enforce bucket cleanup with local testing

It would be ideal to throw a compiler error in the drop implementation 
of AwsLocal that would only get included if the trait was used.

The best we can do right now is panic at the end of a test if the Drop 
trait is called automatically. 

https://doc.rust-lang.org/std/mem/struct.ManuallyDrop.html
bypasses the Drop trait.

* Better implementation of drop gaurd

The previous version used `std::mem::ManuallyDrop` which achieved the 
goal.

However, it had the unintended side effect not dropping the internal 
`BucketClient`. Which we absolutely want to clean up.

This version is a bit simpler, it simply flips a toggle of wether or
not we want to panic when the test ends.