FileType enum macro generation
To add a new file type we had to define:
- a constant
- an enum variant
- mapping enum variant to constant
- mapping constant to enum variant

The macro `make_string_mapped_enum!` generates all of this for us.
Now, we can simply add a single line and get the rest for free.

Also removed the reference to `file_store::Error` that wrapped an `io::Error` 
in lieu of a custom parse error type.