Fix intermittent "dispatch task is gone" in tests
Clients were being created with the same credentials for all tests.

Because the `CLIENT_CACHE` is `static` it is shared for the duration of 
all tests being run. So every test would have the same client. When a
tests finished and dropped it’s client while other tests were still in
flight, the dispatcher would no longer exist and we would get an error.

This fix causes every test to get a fresh client.