Maybe remove appended `.` from prefix when setting start_after
`_list_files` was broken out to be able to specifically test the 
functionality of `.set_start_after()`.

Previously, the tests were passing because of the later filters on
timestamp after a returned s3 object was turned into a `FileInfo`.

The entire history of files was being returned and filtered as a side
effect.